# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar submissões de top 20 jogadores",
        "title": "Top20 API",
        "contact": {
            "name": "API Support",
            "email": "support@top20.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/api/health": {
            "get": {
                "description": "Verifica o status da API e conexão com banco de dados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/players/stats": {
            "get": {
                "description": "Retorna estatísticas de quantas vezes um jogador apareceu em cada posição",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Estatísticas de um jogador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do jogador",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PlayerStats"
                        }
                    },
                    "400": {
                        "description": "Nome do jogador não fornecido",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Jogador não encontrado em nenhuma submissão",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/submissions": {
            "get": {
                "description": "Retorna todas as submissões ou filtra por submittedBy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Listar submissões",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por quem submeteu",
                        "name": "submittedBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Submission"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Submete uma lista com exatamente 20 jogadores. Apenas uma submissão por IP é permitida.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Submeter top 20 jogadores",
                "parameters": [
                    {
                        "description": "Lista de 20 jogadores",
                        "name": "submission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Submissão criada com sucesso"
                    },
                    "400": {
                        "description": "Número incorreto de jogadores ou campos faltando",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "IP já submeteu anteriormente",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.Player": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                }
            }
        },
        "main.PlayerStats": {
            "type": "object",
            "properties": {
                "playerName": {
                    "type": "string"
                },
                "positionBreakdown": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.PositionCount"
                    }
                },
                "totalSubmissions": {
                    "type": "integer"
                }
            }
        },
        "main.PositionCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                }
            }
        },
        "main.Submission": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ipAddress": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Player"
                    }
                },
                "submittedBy": {
                    "type": "string"
                }
            }
        },
        "main.SubmissionRequest": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Player"
                    }
                },
                "submittedBy": {
                    "type": "string"
                }
            }
        }
    }
}